// rest_api_v2/src/command/dto/flying-dragon-response.dto.ts
export class FlyingDragonResponseDto {
  row: number;
  currencycode: string;
  freespinwin: string;
  associatedgoldenchipswins: string;
  autoplaybets: string;
  autoplayprogressivebets: string;
  autoplayproressivewins: string;
  autoplaywins: string;
  avgbets: string;
  bonusbets: string;
  bonusmoneylivegametips: string;
  bonuswins: string;
  comps: string;
  daysplayed: string;
  deductedgoldenchip: string;
  freespinbet: string;
  freespincount: string;
  netpayout: string;
  gincomeshare: string;
  gamescnt: string;
  goldenchipbet: string;
  goldenchipwin: string;
  overallgamepayout: string;
  playerscnt: string;
  jackpotbets: string;
  jackpotwins: string;
  real_hold: string;
  realbets: string;
  realmoneyincome: string;
  realmoneylivegametips: string;
  realpayout: string;
  realjackpotwins: string;
  realwins: string;
  returnedgoldenchips: string;
  totalbets: string;
  totalgamebets: string;
  totalgamewins: string;
  totalincome: string;
  refund: string;
  totalplayerwins: string;
  company: string;
  casino: string;
  date1: string;
  week1: string;
  month1: string;
  playercurrency: string;
  language: string;
  clienttype: string;
  clientplatform: string;
  country: string;
  tlentity: string;
  gametype: string;
  gamename: string;
  maximumrtp: string;
  playercode: string;
  kiosk: string;
  username: string;
}